+++ src/contract-testing/consumer/amf-ais-consumer.pact.test.ts
@@ -1,18 +1,14 @@
 import { MessageConsumerPact, synchronousBodyHandler } from '@pact-foundation/pact';
-import { COMPONENT_ID } from '../../data-types/constants';
 import { string } from '@pact-foundation/pact/src/v3/matchers';
-import pkg from "@pact-foundation/pact-node";
-const { publishPacts } = pkg;
-
 import path from 'node:path';
 
 describe('AMF & AIS - Contract Testing - Consumer', () => {
   const messagePact = new MessageConsumerPact({
-    consumer: COMPONENT_ID,
+    consumer: 'AccountInterventionsServiceConsumer',
     dir: path.resolve(process.cwd(), 'pacts'),
     pactfileWriteMode: 'update',
-    provider: 'Account Management Frontend',
-    logLevel: "info",
+    provider: 'AccountManagementFrontendProvider',
+    logLevel: "error",
   });
 
   describe('Incoming delete account event from Account Management Frontend', () => {
@@ -35,25 +31,3 @@ function syncMessageHandler(event: any) {
   }
   return;
 }
-
-const publishPact = async () => {
-  console.log("STARTING PUBLISH PACT");
-  try {
-    const publishOptions = {
-      pactFilesOrDirs: [path.resolve(process.cwd(), "pacts")],
-      pactBroker: process.env['PACT_BROKER_URL']!,
-      pactBrokerUsername: process.env['PACT_BROKER_USER']!,
-      pactBrokerPassword: process.env['PACT_BROKER_PASSWORD']!,
-      logLevel: "info",
-      consumerVersion: process.env['CONSUMER_APP_VERSION']!,
-      branch: "main"
-    };
-
-    await publishPacts(publishOptions);
-  } catch (err) {
-    console.error("UNABLE TO PUBLISH PACTS ", err);
-    process.exitCode = 1;
-  }
-};
-
-void publishPact();
